+++ js-web/vue-vite/src/WebTemplate/views/Callback.vue
@@ -17,6 +17,7 @@ export default {
     const hash = ref('No Hash Found')
 
     onMounted(async () => {
+      // NOSONAR const { hash, href, port, hostname, protocol, ...etc } = window.location
       console.log('Callback mounted!', route.hash, route) // deal with hashes here if necessary
       hash.value = route.hash.substring(1) // <access_token>;<refresh_token>;<groups JSON string>
       // verify first, if ok, do login, else send to forbidden // split(';')